inference/requirements.txt
tensorflow
fastapi[all]
Pillow
python-multipart
gradio==4.8.0
pymongo